init_ws.sh: Move configuration process to Yocto
As machine confgurations are moved to the Yocto layers and are not set
using config fragments in gyroidos_build, the functionality for copying
config fragments to local.conf is removed. Further configurations in
local.conf are now made using regualar expressions instead of appending
to make the configuration process idempotent and robust against double
initialisation.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>