Jenkinsfile: create initial Version
Add a Jenkinsfile for building pull requests on this repo. Building is
performed by calling the generic pipeline in gyroidos manifest. If a
pull request is built, the PR branch is passed to the generic pipeline
and considered automatically. Additional pull request branches can be
added via the PR_BRANCHES parameter in the Jenkins interface.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>