provisioning guestos_config_creator: do not generate empty verity hash
In case the the config paramter for root_hash is not set an empty string
is used as defaut. This leads to an empty image_verity_sha256 in the
final protobuf-text config file. This is now checked so the image_vertity
field will only be generated in the final config file if a root_hash was
set.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>