yocto/copy_image_to_disk: allow flexible partition layout
Instead of defining PARTNUM fixed to 2, we now readout the
partition number from the gyroidoimage.img. Assuming the partition
labeled with gyroidos is last. This allows to us to use the
copy_image_to_disk.sh script also for, e.g., the beaglev-fire image
which in addition has a special boot partition as second partition.

Further, also use the bmap-tool if available, similar to the
copy_image_to_disk_mbr.sh.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>